+++ CHANGES.rst
@@ -23,6 +23,7 @@ Unreleased
 **Removed**
 
 * Drop official support for Python 2.7.
+* ``Multireddit.rename()`` no longer works due to a change in the Reddit API.
 
 6.3.1 (2019/06/10)
 ------------------